test: rename Unmock.ts to unmock.ts
