+++ test/unit/src/util/command_util.test.ts
@@ -3,7 +3,7 @@ import CommandUtil from "../../../../src/util/command_util";
 import { CommandScript } from "../../../../src/command/command_script";
 import TokenisedCommand from "../../../../src/dto/tokenised_command";
 import TerminalUtil from "../../../../src/util/terminal_util";
-import { unmock } from "../../helper/Unmock";
+import { unmock } from "../../helper/unmock";
 import MetaImportUtil from "../../../../src/util/meta_import_util";
 
 describe("CommandUtil", () => {
